Split native Zenmap NSE script database helpers
